Adding a test with synthetic data

+++ ino_tree3/decision_forest.h
@@ -1,2 +1,2 @@
-// unsigned char evaluate_forest(float* args);
-unsigned char evaluate_forest(long* args);
+unsigned char evaluate_forest(float* args);
+// unsigned char evaluate_forest(long* args);
